# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SenAgri B2B",
        "description": "Documentation officielle de l'API SenAgri B2B",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/offres": {
            "get": {
                "tags": [
                    "Offres"
                ],
                "summary": "Lister les offres",
                "operationId": "05bfdbaa2b1418bc0055ac118ad7bceb",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "description": "Recherche par nom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categorie_id",
                        "in": "query",
                        "description": "Filtrer par catégorie",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "prix_min",
                        "in": "query",
                        "description": "Prix minimum",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "prix_max",
                        "in": "query",
                        "description": "Prix maximum",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "disponible",
                        "in": "query",
                        "description": "Disponibilité",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "tri",
                        "in": "query",
                        "description": "Tri (prix_asc, prix_desc, ancien, nom)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des offres"
                    }
                }
            },
            "post": {
                "tags": [
                    "Offres"
                ],
                "summary": "Créer une offre",
                "operationId": "53929be4637f259e9861ff550a140019",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categorie_id",
                                    "nom",
                                    "description",
                                    "prix_unitaire",
                                    "quantite_disponible",
                                    "unite"
                                ],
                                "properties": {
                                    "categorie_id": {
                                        "type": "integer"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "prix_unitaire": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "quantite_disponible": {
                                        "type": "integer"
                                    },
                                    "unite": {
                                        "type": "string"
                                    },
                                    "est_disponible": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre créée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/offres/{id}": {
            "get": {
                "tags": [
                    "Offres"
                ],
                "summary": "Afficher une offre",
                "operationId": "aa918bd8526044eeffcb7c8f6944bfd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail d'une offre"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Offres"
                ],
                "summary": "Modifier une offre",
                "operationId": "34b83097859cea2fca183a26d0233e0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "categorie_id": {
                                        "type": "integer"
                                    },
                                    "nom": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "prix_unitaire": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "quantite_disponible": {
                                        "type": "integer"
                                    },
                                    "unite": {
                                        "type": "string"
                                    },
                                    "est_disponible": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offre mise à jour"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Offres"
                ],
                "summary": "Supprimer une offre",
                "operationId": "31370150957657a8670d33b9bf9fd5c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre supprimée"
                    },
                    "404": {
                        "description": "Offre introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}